Crate rusoto_guardduty

Amazon GuardDuty is a continuous security monitoring service that analyzes and processes the following data sources: VPC Flow Logs, AWS CloudTrail event logs, and DNS logs. It uses threat intelligence feeds, such as lists of malicious IPs and domains, and machine learning to identify unexpected and potentially unauthorized and malicious activity within your AWS environment. This can include issues like escalations of privileges, uses of exposed credentials, or communication with malicious IPs, URLs, or domains. For example, GuardDuty can detect compromised EC2 instances serving malware or mining bitcoin. It also monitors AWS account access behavior for signs of compromise, such as unauthorized infrastructure deployments, like instances deployed in a region that has never been used, or unusual API calls, like a password policy change to reduce password strength. GuardDuty informs you of the status of your AWS environment by producing security findings that you can view in the GuardDuty console or through Amazon CloudWatch events. For more information, see Amazon GuardDuty User Guide.

If you're using the service, you're probably looking for GuardDutyClient and GuardDuty.

Structs

AcceptInvitationRequest

AcceptInvitationResponse

AccessKeyDetails

AccountDetail

Action

ArchiveFindingsRequest

ArchiveFindingsResponse

AwsApiCallAction

City

Condition

Country

CreateDetectorRequest

CreateDetectorResponse

CreateFilterRequest

CreateFilterResponse

CreateIPSetRequest

CreateIPSetResponse

CreateMembersRequest

CreateMembersResponse

CreateSampleFindingsRequest

CreateSampleFindingsResponse

CreateThreatIntelSetRequest

CreateThreatIntelSetResponse

DeclineInvitationsRequest

DeclineInvitationsResponse

DeleteDetectorRequest

DeleteDetectorResponse

DeleteFilterRequest

DeleteFilterResponse

DeleteIPSetRequest

DeleteIPSetResponse

DeleteInvitationsRequest

DeleteInvitationsResponse

DeleteMembersRequest

DeleteMembersResponse

DeleteThreatIntelSetRequest

DeleteThreatIntelSetResponse

DisassociateFromMasterAccountRequest

DisassociateFromMasterAccountResponse

DisassociateMembersRequest

DisassociateMembersResponse

DnsRequestAction

DomainDetails

Finding

FindingCriteria

FindingStatistics

GeoLocation

GetDetectorRequest

GetDetectorResponse

GetFilterRequest

GetFilterResponse

GetFindingsRequest

GetFindingsResponse

GetFindingsStatisticsRequest

GetFindingsStatisticsResponse

GetIPSetRequest

GetIPSetResponse

GetInvitationsCountRequest

GetInvitationsCountResponse

GetMasterAccountRequest

GetMasterAccountResponse

GetMembersRequest

GetMembersResponse

GetThreatIntelSetRequest

GetThreatIntelSetResponse

GuardDutyClient
A client for the Amazon GuardDuty API.

IamInstanceProfile

InstanceDetails

Invitation

InviteMembersRequest

InviteMembersResponse

ListDetectorsRequest

ListDetectorsResponse

ListFiltersRequest

ListFiltersResponse

ListFindingsRequest

ListFindingsResponse

ListIPSetsRequest

ListIPSetsResponse

ListInvitationsRequest

ListInvitationsResponse

ListMembersRequest

ListMembersResponse

ListThreatIntelSetsRequest

ListThreatIntelSetsResponse

LocalPortDetails

Master

Member

NetworkConnectionAction

NetworkInterface

Organization

PortProbeAction

PortProbeDetail

PrivateIpAddressDetails

ProductCode

RemoteIpDetails

RemotePortDetails

Resource

SecurityGroup

Service

SortCriteria

StartMonitoringMembersRequest

StartMonitoringMembersResponse

StopMonitoringMembersRequest

StopMonitoringMembersResponse

Tag

UnarchiveFindingsRequest

UnarchiveFindingsResponse

UnprocessedAccount

UpdateDetectorRequest

UpdateDetectorResponse

UpdateFilterRequest

UpdateFilterResponse

UpdateFindingsFeedbackRequest

UpdateFindingsFeedbackResponse

UpdateIPSetRequest

UpdateIPSetResponse

UpdateThreatIntelSetRequest

UpdateThreatIntelSetResponse

Enums

AcceptInvitationError
Errors returned by AcceptInvitation

ArchiveFindingsError
Errors returned by ArchiveFindings

CreateDetectorError
Errors returned by CreateDetector

CreateFilterError
Errors returned by CreateFilter

CreateIPSetError
Errors returned by CreateIPSet

CreateMembersError
Errors returned by CreateMembers

CreateSampleFindingsError
Errors returned by CreateSampleFindings

CreateThreatIntelSetError
Errors returned by CreateThreatIntelSet

DeclineInvitationsError
Errors returned by DeclineInvitations

DeleteDetectorError
Errors returned by DeleteDetector

DeleteFilterError
Errors returned by DeleteFilter

DeleteIPSetError
Errors returned by DeleteIPSet

DeleteInvitationsError
Errors returned by DeleteInvitations

DeleteMembersError
Errors returned by DeleteMembers

DeleteThreatIntelSetError
Errors returned by DeleteThreatIntelSet

DisassociateFromMasterAccountError
Errors returned by DisassociateFromMasterAccount

DisassociateMembersError
Errors returned by DisassociateMembers

GetDetectorError
Errors returned by GetDetector

GetFilterError
Errors returned by GetFilter

GetFindingsError
Errors returned by GetFindings

GetFindingsStatisticsError
Errors returned by GetFindingsStatistics

GetIPSetError
Errors returned by GetIPSet

GetInvitationsCountError
Errors returned by GetInvitationsCount

GetMasterAccountError
Errors returned by GetMasterAccount

GetMembersError
Errors returned by GetMembers

GetThreatIntelSetError
Errors returned by GetThreatIntelSet

InviteMembersError
Errors returned by InviteMembers

ListDetectorsError
Errors returned by ListDetectors

ListFiltersError
Errors returned by ListFilters

ListFindingsError
Errors returned by ListFindings

ListIPSetsError
Errors returned by ListIPSets

ListInvitationsError
Errors returned by ListInvitations

ListMembersError
Errors returned by ListMembers

ListThreatIntelSetsError
Errors returned by ListThreatIntelSets

StartMonitoringMembersError
Errors returned by StartMonitoringMembers

StopMonitoringMembersError
Errors returned by StopMonitoringMembers

UnarchiveFindingsError
Errors returned by UnarchiveFindings

UpdateDetectorError
Errors returned by UpdateDetector

UpdateFilterError
Errors returned by UpdateFilter

UpdateFindingsFeedbackError
Errors returned by UpdateFindingsFeedback

UpdateIPSetError
Errors returned by UpdateIPSet

UpdateThreatIntelSetError
Errors returned by UpdateThreatIntelSet

Traits

GuardDuty
Trait representing the capabilities of the Amazon GuardDuty API. Amazon GuardDuty clients implement this trait.