Struct rusoto_guardduty::GuardDutyClient

pub struct GuardDutyClient { /* fields omitted */ }

A client for the Amazon GuardDuty API.

Methods

impl GuardDutyClient

pub fn new(region: Region) -> GuardDutyClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> GuardDutyClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,